+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+- Adding document formatting support (Issue #23)
+- New configuration setting: raku.formattingEnabled
+
 ## 0.3.1 2024-02-17
 
 - Adding this changelog